5. Project Files Submission and Documentation/requirements.txt
streamlit==1.30.0
google-generativeai==0.3.2
python-dotenv==1.0.0
